requirements.txt
asgiref==3.3.4
Django==3.1.6
django-crispy-forms==1.12.0
django-filter==2.4.0
Pillow==8.2.0
pytz==2021.1
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.5
